requirements.txt
flask
flask-cors
pandas
google-genai
gunicorn